refactor: move positioning logic to utils and reduce number of re-renders
